refactor: modularize slash command and quick selector
