+++ src/components/prompts/blocks/quick-selector/index.ts
@@ -1,6 +1,9 @@
 // src/components/prompts/quick-selector/index.ts
 export { QuickBlockSelector } from './QuickBlockSelector';
 export { BlockItem } from './BlockItem';
+export { useBlocks } from './useBlocks';
+export { useBlockInsertion } from './useBlockInsertion';
+export { calculateDropdownPosition } from './positionUtils';
 export type { QuickBlockSelectorProps, BlockItemProps } from './types';
 
 // Type definitions file: src/components/prompts/quick-selector/types.ts
